Finished!
Added args[] implementation to main method and added actual info to
Help.help(). I might update it once or twice to add additional features
or more secure keyEncryption(If its not secure already).

In order to use it the "CRYPT_KEYS" environment variable needs to be
set, and the project should be converted to an .exe file with its
location set in the "Path" environment variable.